+++ CHANGELOG.md
@@ -6,7 +6,25 @@ There are next changes:
 
 - dev dependence vaimo/composer-patches was updated from 3.4.3 to 3.23.1;
 - patch level for patches was provided;
-- phpunit6 support was added.
+- phpunit6 support was added;
+- class static protected constant was fixed;
+- class constants inheritance was fixed:
+  - before fix:
+    ```php
+    class A {const NAME = 'A';}
+    class B {}
+    \Badoo\SoftMocks::redefineConstant(A::class . '::NAME', 'B');
+    echo A::NAME . "\n"; // B
+    echo B::NAME . "\n"; // A
+    ```
+  - after fix:
+    ```php
+    class A {const NAME = 'A';}
+    class B {}
+    \Badoo\SoftMocks::redefineConstant(A::class . '::NAME', 'B');
+    echo A::NAME . "\n"; // B
+    echo B::NAME . "\n"; // B
+    ```
 
 ## v1.3.5